alert-generation/requirements.txt
quixstreams<2.10
python-dotenv
geopy
requests